running MOST from the loop
