+++ ercot/case8/dsostub/.gitignore
@@ -5,3 +5,4 @@
 dam.json
 *.txt
 test*.m
+dam*.m
