+++ ercot/case8/dsostub/clean.sh
@@ -7,4 +7,5 @@ rm *dict.json
 rm *.dat
 rm *.txt
 rm dam.json
+rm dam*.m
 rm test_dam*.m
