--- ercot/case8/dsostub/dam_loads3.m
@@ -1,29 +1,0 @@
-## Author: tom <tom@uvm>
-## Created: 2020-10-02
-function loadprofile = dam_loads3
-%% define constants
-[CT_LABEL, CT_PROB, CT_TABLE, CT_TBUS, CT_TGEN, CT_TBRCH, CT_TAREABUS, ...
-    CT_TAREAGEN, CT_TAREABRCH, CT_ROW, CT_COL, CT_CHGTYPE, CT_REP, ...
-    CT_REL, CT_ADD, CT_NEWVAL, CT_TLOAD, CT_TAREALOAD, CT_LOAD_ALL_PQ, ...
-    CT_LOAD_FIX_PQ, CT_LOAD_DIS_PQ, CT_LOAD_ALL_P, CT_LOAD_FIX_P, ...
-    CT_LOAD_DIS_P, CT_TGENCOST, CT_TAREAGENCOST, CT_MODCOST_F, ...
-    CT_MODCOST_X] = idx_ct;
-
-loadprofile = struct( ...
-    'type', 'mpcData', ...
-    'table', CT_TLOAD, ...
-    'rows', [1,2,3,4,5,6,7,8], ...
-    'col', CT_LOAD_DIS_P, ...
-    'chgtype', CT_REP, ...
-    'values', [] );
-
-loadprofile.values(:, 1, 1) = [8000.0; 8000.0; 9000.0; 8000.0];
-loadprofile.values(:, 1, 2) = [3625.0; 3625.0; 3625.0; 3625.0];
-loadprofile.values(:, 1, 3) = [4625.0; 4635.0; 4625.0; 4625.0];
-loadprofile.values(:, 1, 4) = [3625.0; 3625.0; 3625.0; 3625.0];
-loadprofile.values(:, 1, 5) = [3625.0; 3625.0; 3625.0; 3825.0];
-loadprofile.values(:, 1, 6) = [3625.0; 3625.0; 3625.0; 3625.0];
-loadprofile.values(:, 1, 7) = [3625.0; 3625.0; 3625.0; 3625.0];
-loadprofile.values(:, 1, 8) = [3625.0; 3625.0; 3625.0; 3625.0];
-
-endfunction
