--- ercot/case8/dsostub/dam_loads4.m
@@ -1,31 +1,0 @@
-## Author: tom <tom@uvm>
-## Created: 2020-10-02
-function loadprofile = dam_loads4
-%% define constants
-[PQ, PV, REF, NONE, BUS_I, BUS_TYPE, PD, QD, GS, BS, BUS_AREA, VM, ...
-   VA, BASE_KV, ZONE, VMAX, VMIN, LAM_P, LAM_Q, MU_VMAX, MU_VMIN] = idx_bus;
-[CT_LABEL, CT_PROB, CT_TABLE, CT_TBUS, CT_TGEN, CT_TBRCH, CT_TAREABUS, ...
-    CT_TAREAGEN, CT_TAREABRCH, CT_ROW, CT_COL, CT_CHGTYPE, CT_REP, ...
-    CT_REL, CT_ADD, CT_NEWVAL, CT_TLOAD, CT_TAREALOAD, CT_LOAD_ALL_PQ, ...
-    CT_LOAD_FIX_PQ, CT_LOAD_DIS_PQ, CT_LOAD_ALL_P, CT_LOAD_FIX_P, ...
-    CT_LOAD_DIS_P, CT_TGENCOST, CT_TAREAGENCOST, CT_MODCOST_F, ...
-    CT_MODCOST_X] = idx_ct;
-
-loadprofile = struct( ...
-    'type', 'mpcData', ...
-    'table', CT_TBUS, ...
-    'rows', [1,2,3,4,5,6,7,8], ...
-    'col', PD, ...
-    'chgtype', CT_REP, ...
-    'values', [] );
-
-loadprofile.values(:, 1, 1) = [2000.0; 2000.0; 2000.0; 2000.0];
-loadprofile.values(:, 1, 2) = [2000.0; 2000.0; 2000.0; 2000.0];
-loadprofile.values(:, 1, 3) = [2000.0; 3000.0; 2000.0; 2000.0];
-loadprofile.values(:, 1, 4) = [2000.0; 2000.0; 2000.0; 2800.0];
-loadprofile.values(:, 1, 5) = [1000.0; 1000.0; 1000.0; 1000.0];
-loadprofile.values(:, 1, 6) = [1000.0; 1000.0; 1000.0; 1000.0];
-loadprofile.values(:, 1, 7) = [ 500.0;  500.0;  500.0;  500.0];
-loadprofile.values(:, 1, 8) = [  10.0;   10.0;   10.0;   10.0];
-
-endfunction
