--- ercot/case8/dsostub/dam_ucdata.m
@@ -1,33 +1,0 @@
-function [xgd_table] = dam_ucdata (mpc)
-xgd_table.colnames = {
-    'CommitKey', ...
-};
-xgd_table.data = [
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-    1;
-   -1;
-   -1;
-   -1;
-   -1;
-   -1;
-    2;
-    2;
-    2;
-    2;
-    2;
-    2;
-    2;
-    2;
-];
-endfunction
