+++ src/tesp_support/tesp_support/api.py
@@ -63,6 +63,8 @@ from .run_test_case import InitializeTestCaseReports
 
 from .prep_eplus import make_gld_eplus_case
 
+from .parse_msout import read_most_solution
+
 #from .process_agents import process_agents
 #from .process_eplus import process_eplus
 #from .process_gld import process_gld
